Extract non `ink!` specific drink API for use with `solang`
