+++ src/lib.rs
@@ -4,6 +4,8 @@ pub mod drink_api;
 
 pub mod ink_build;
 
+pub mod ink_drink;
+
 #[cfg(feature = "riscv")]
 pub use drink_riscv as drink;
 #[cfg(feature = "wasm")]
